Add an abstraction for rational numbers and fix related type conversion inconsistencies
All potential problems with the rational numbers come from the
fact that FFmpeg represents them using C int type which may not
be equivalent to Rust's i32 on all HW platforms. In addition,
even though negative time bases do not make any sense, representing
them using a u32 pair isn't a very good idea since FFmpeg represents
them using an int pair and the type conversions may lead to unexpected
results.

+++ ac-ffmpeg/src/codec/audio/resampler.rs
@@ -220,7 +220,7 @@ impl AudioResampler {
             ));
         }
 
-        let frame = frame.with_time_base(TimeBase::new(1, self.source_sample_rate));
+        let frame = frame.with_time_base(TimeBase::new(1, self.source_sample_rate as i32));
 
         unsafe {
             match ffw_audio_resampler_push_frame(self.ptr, frame.as_ptr()) {
@@ -259,7 +259,7 @@ impl AudioResampler {
     pub fn take(&mut self) -> Result<Option<AudioFrame>, Error> {
         let mut fptr = ptr::null_mut();
 
-        let tb = TimeBase::new(1, self.target_sample_rate);
+        let tb = TimeBase::new(1, self.target_sample_rate as i32);
 
         unsafe {
             match ffw_audio_resampler_take_frame(self.ptr, &mut fptr) {
